@@ -37,9 +37,10 @@ dependencies:
   google_fonts: ^6.0.0
   flutter_spinkit: ^5.2.0
   webview_flutter: ^4.8.2
-  youtube_player_flutter: ^8.1.2
+  youtube_player_flutter: 9.1.3
   cached_network_image: ^3.3.0
   shimmer: ^3.0.0
+
   # Navigation
   go_router: ^10.0.0
 
